server/requirements.txt
absl-py
amqp
asgiref
astunparse
billiard
celery
certifi
cffi
channels
channels_redis
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
colorama
cryptography
Django
django-celery-results
django-cors-headers
django-rest-framework
djangorestframework
djangorestframework-simplejwt
flatbuffers
flower
gast
google-pasta
grpcio
h5py
humanize
idna
joblib
keras
kombu
libclang
Markdown
markdown-it-py
MarkupSafe
mdurl
ml-dtypes
msgpack
mysqlclient
namex
numpy
opencv-python
opt_einsum
optree
packaging
pandas
pillow
pip
prometheus_client
prompt_toolkit
protobuf
pycparser
Pygments
PyJWT
python-dateutil
python-dotenv
pytz
redis
requests
rich
scikit-learn
scipy
setuptools
six
sqlparse
tensorboard
tensorboard-data-server
tensorflow
tensorflow_intel
termcolor
threadpoolctl
tornado
typing_extensions
tzdata
urllib3
vine
wcwidth
Werkzeug
wheel
wrapt
cryptography
Django
djangorestframework
djangorestframework_simplejwt
numpy
opencv_python
pandas
Pillow
PyJWT
python-dotenv
scikit_learn
tensorflow
tensorflow_intel